Add VSCode settings for Python testing
This commit adds a new ".vscode/settings.json" file with configurations for Python testing in VSCode. The "pytestArgs" option is set to run tests from the "tests" directory. Additionally, the "unittestEnabled" option is disabled, while "pytestEnabled" is set to true. These settings allow for seamless integration of pytest for running Python tests within the VSCode environment.