feat: add zip ignore pattern for deployment
This commit adds the configuration to enable pytest in the VS Code
settings. Additionally, a new "appService.zipIgnorePattern" setting is
introduced, which specifies the patterns to exclude while creating a zip file
for deployment. These changes aim to improve the testing experience and
optimize the deployment process by excluding unnecessary files and
directories from the deployment package.